Introduce variable pax in struct switchtec_dev
Add variable pax for PAX id, which is only valid for the Switchtec
PAX variants. The PAX id is encoded in the command option field when
issue command.

With valid PAX id, the firmware either execute the command on local
switch if the PAX id is for local, or forward the command to remote
PAX switch. In the remote case, local firmware acts as a command agent.
The actual execution of the command is on the remote PAX.

The PAX ID can be set by the user using switchtec_set_pax_id().